requirements.txt
# Core Dependencies
python-dotenv==1.0.0
google-generativeai==0.3.0

# LLM & AI Framework
langgraph==0.0.21
langchain==0.1.10
langchain-core==0.1.23
langchain-google-genai==0.0.5

# Utilities
requests==2.31.0